Assignment 8 - FINAL BOSS/cli/requirements.txt
boto3
requests
rich
pyfiglet
python-dotenv
